+++ chain/network/src/peer_manager/mod.rs
@@ -1,2 +1,3 @@
+pub(crate) mod db_types;
 pub(crate) mod peer_manager_actor;
 pub(crate) mod peer_store;
