+++ chain/network/src/peer_manager/peer_manager_actor.rs
@@ -1,5 +1,6 @@
 use crate::peer::codec::Codec;
 use crate::peer::peer_actor::PeerActor;
+use crate::peer_manager::db_types::{KnownPeerState, KnownPeerStatus};
 use crate::peer_manager::peer_store::{PeerStore, TrustLevel};
 use crate::routing::edge_validator_actor::EdgeValidatorHelper;
 #[cfg(all(
@@ -32,11 +33,10 @@ use actix::{
 use futures::task::Poll;
 use futures::{future, Stream, StreamExt};
 use near_network_primitives::types::{
-    AccountOrPeerIdOrHash, Ban, BlockedPorts, InboundTcpConnect, KnownPeerState, KnownPeerStatus,
-    KnownProducer, NetworkConfig, NetworkViewClientMessages, NetworkViewClientResponses,
-    OutboundTcpConnect, PeerIdOrHash, PeerManagerRequest, PeerType, Ping, Pong, QueryPeerStats,
-    RawRoutedMessage, ReasonForBan, RoutedMessage, RoutedMessageBody, RoutedMessageFrom,
-    StateResponseInfo,
+    AccountOrPeerIdOrHash, Ban, BlockedPorts, InboundTcpConnect, KnownProducer, NetworkConfig,
+    NetworkViewClientMessages, NetworkViewClientResponses, OutboundTcpConnect, PeerIdOrHash,
+    PeerManagerRequest, PeerType, Ping, Pong, QueryPeerStats, RawRoutedMessage, ReasonForBan,
+    RoutedMessage, RoutedMessageBody, RoutedMessageFrom, StateResponseInfo,
 };
 use near_performance_metrics::framed_write::FramedWrite;
 use near_performance_metrics_macros::perf;
